moved search from uid to core and added nav button for search
